+++ lib/article_json.rb
@@ -47,6 +47,7 @@ require_relative 'article_json/export/common/html/elements/heading'
 require_relative 'article_json/export/common/html/elements/paragraph'
 require_relative 'article_json/export/common/html/elements/list'
 require_relative 'article_json/export/common/html/elements/quote'
+require_relative 'article_json/export/common/html/elements/text_box'
 
 require_relative 'article_json/export/html/elements/base'
 require_relative 'article_json/export/html/elements/text'
