+++ lib/article_json/export/amp/elements/text_box.rb
@@ -3,23 +3,7 @@ module ArticleJSON
     module AMP
       module Elements
         class TextBox < Base
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Float
-
-          # Generate a `<div>` node containing all text box elements
-          # return [Nokogiri::XML::NodeSet]
-          def export
-            create_element(:div, node_opts) do |div|
-              @element.content.each do |child_element|
-                div.add_child(Base.new(child_element).export)
-              end
-            end
-          end
-
-          private
-
-          def node_opts
-            { class: ['text-box', floating_class].compact.join(' ') }
-          end
+          include ArticleJSON::Export::Common::HTML::Elements::TextBox
         end
       end
     end
